week_06/day_01/Intro_Bear_examples/src/main/java/Bear.java
public class Bear {

    private double weight;
    private String name;
    private int age;

    public Bear(String name, int age, double weight) {
        this.name = name;
        this.age = age;
        this.weight = weight;
    }

    public String getName() {
        return this.name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getAge() {
        return this.age;
    }

    public double getWeight() {
        return this.weight;
    }

    public boolean readyToHibernate() {
        return this.weight >= 80.00;
    }
}





